WIP: revert and refactor files
