+++ consensus/tendermint/core/core.go
@@ -25,18 +25,11 @@ import (
 	"time"
 
 	"github.com/clearmatics/autonity/common"
-	"github.com/clearmatics/autonity/consensus"
 	"github.com/clearmatics/autonity/consensus/tendermint"
 	"github.com/clearmatics/autonity/core/types"
 	"github.com/clearmatics/autonity/event"
 	"github.com/clearmatics/autonity/log"
 	"gopkg.in/karalabe/cookiejar.v2/collections/prque"
-)
-
-const (
-	initialProposeTimeout = 5 * time.Second
-	//initialPrevoteTimeout   = 5 * time.Second
-	//initialPrecommitTimeout = 5 * time.Second
 )
 
 var (
@@ -72,15 +65,10 @@ var (
 	errNilPrevoteSent = errors.New("timer expired and nil prevote sent")
 )
 
-var (
-	// msgPriority is defined for calculating processing priority to speedup consensus
-	// msgProposal > msgPrecommit > msgPrevote
-	msgPriority = map[uint64]int{
-		msgProposal:  1,
-		msgPrecommit: 2,
-		msgPrevote:   3,
-	}
-)
+type Engine interface {
+	Start() error
+	Stop() error
+}
 
 // TODO: add better logging
 // New creates an Istanbul consensus core
@@ -103,8 +91,6 @@ func New(backend tendermint.Backend, config *tendermint.Config) Engine {
 	c.validateFn = c.checkValidatorSignature
 	return c
 }
-
-// ----------------------------------------------------------------------------
 
 type core struct {
 	config  *tendermint.Config
@@ -326,781 +312,4 @@ func PrepareCommittedSeal(hash common.Hash) []byte {
 	buf.Write(hash.Bytes())
 	buf.Write([]byte{byte(msgPrecommit)})
 	return buf.Bytes()
-}
-
-func (c *core) onTimeoutPropose(h int64, r int64) {
-	if h == c.currentRoundState.Height().Int64() && r == c.currentRoundState.Round().Int64() && c.step == StepAcceptProposal {
-		c.sendPrevote(true)
-		c.setStep(StepProposeDone)
-	}
-}
-
-func (c *core) onTimeoutPrevote(h int64, r int64) {
-	if h == c.currentRoundState.Height().Int64() && r == c.currentRoundState.Round().Int64() && c.step == StepProposeDone {
-		c.sendPrecommit(true)
-		c.setStep(StepPrevoteDone)
-	}
-}
-
-func (c *core) onTimeoutPrecommit(h int64, r int64) {
-	if h == c.currentRoundState.Height().Int64() && r == c.currentRoundState.Round().Int64() {
-		c.startRound(new(big.Int).Add(c.currentRoundState.Height(), common.Big1))
-	}
-}
-
-//---------------------------------------Timeout---------------------------------------
-
-type timeoutEvent struct{}
-
-type timeout struct {
-	timer   *time.Timer
-	started bool
-	sync.RWMutex
-}
-
-// runAfterTimeout() will be run in a separate go routine, so values used inside the function needs to be managed separately
-func (t *timeout) scheduleTimeout(stepTimeout time.Duration, height int64, round int64, runAfterTimeout func(h int64, r int64)) *time.Timer {
-	t.Lock()
-	defer t.Unlock()
-	t.started = true
-	t.timer = time.AfterFunc(stepTimeout, func() {
-		runAfterTimeout(height, round)
-	})
-	return t.timer
-}
-
-func (t *timeout) stopTimer() bool {
-	t.RLock()
-	defer t.RUnlock()
-	return t.timer.Stop()
-}
-
-// The timeout may need to be changed depending on the Step
-func timeoutPropose(round int64) time.Duration {
-	return initialProposeTimeout + time.Duration(round)*time.Second
-}
-
-func timeoutPrevote(round int64) time.Duration {
-	return initialProposeTimeout + time.Duration(round)*time.Second
-}
-
-func timeoutPrecommit(round int64) time.Duration {
-	return initialProposeTimeout + time.Duration(round)*time.Second
-}
-
-//---------------------------------------Handler---------------------------------------
-
-// Start implements core.Engine.Start
-func (c *core) Start() error {
-	// Start a new round from last height + 1
-	c.startRound(common.Big0)
-
-	// Tests will handle events itself, so we have to make subscribeEvents()
-	// be able to call in test.
-	c.subscribeEvents()
-	go c.handleEvents()
-
-	return nil
-}
-
-// Stop implements core.Engine.Stop
-func (c *core) Stop() error {
-	c.stopTimer()
-	c.unsubscribeEvents()
-
-	// Make sure the handler goroutine exits
-	c.handlerStopCh <- struct{}{}
-	return nil
-}
-
-// TODO: update all of the TypeMuxSilent to event.Feed
-
-// Subscribe both internal and external events
-func (c *core) subscribeEvents() {
-	c.events = c.backend.EventMux().Subscribe(
-		// external events
-		tendermint.NewUnminedBlockEvent{},
-		tendermint.MessageEvent{},
-		// internal events
-		backlogEvent{},
-	)
-	// TODO: not sure why a backend EventMux is being used for core internal events, lazy coding
-	c.timeoutSub = c.backend.EventMux().Subscribe(
-		timeoutEvent{},
-	)
-	// TODO: not sure why a backend EventMux is being used for core internal events, lazy coding
-	c.finalCommittedSub = c.backend.EventMux().Subscribe(
-		tendermint.CommitEvent{},
-	)
-}
-
-// Unsubscribe all events
-func (c *core) unsubscribeEvents() {
-	c.events.Unsubscribe()
-	c.timeoutSub.Unsubscribe()
-	c.finalCommittedSub.Unsubscribe()
-}
-
-func (c *core) handleEvents() {
-	// Clear step
-	defer func() {
-		c.currentRoundState = nil
-		<-c.handlerStopCh
-	}()
-
-	for {
-		select {
-		case ev, ok := <-c.events.Chan():
-			if !ok {
-				return
-			}
-			// A real ev arrived, process interesting content
-			switch e := ev.Data.(type) {
-			case tendermint.NewUnminedBlockEvent:
-				pb := &e.NewUnminedBlock
-				err := c.handleUnminedBlock(pb)
-				//TODO: return consensus.ErrFutureBlock and handle it gracefully
-				if err == errFutureHeightMessage {
-					c.storeUnminedBlockMsg(pb)
-				}
-			case tendermint.MessageEvent:
-				if err := c.handleMsg(e.Payload); err == nil {
-					c.backend.Gossip(c.valSet, e.Payload)
-				}
-			case backlogEvent:
-				// No need to check signature for internal messages
-				if err := c.handleCheckedMsg(e.msg, e.src); err == nil {
-					p, err := e.msg.Payload()
-					if err != nil {
-						c.logger.Warn("Get message payload failed", "err", err)
-						continue
-					}
-					c.backend.Gossip(c.valSet, p)
-				}
-			}
-		case _, ok := <-c.timeoutSub.Chan():
-			if !ok {
-				return
-			}
-			c.handleTimeoutMsg()
-		case ev, ok := <-c.finalCommittedSub.Chan():
-			if !ok {
-				return
-			}
-			switch ev.Data.(type) {
-			case tendermint.CommitEvent:
-				c.handleCommit()
-			}
-		}
-	}
-}
-
-// sendEvent sends events to mux
-func (c *core) sendEvent(ev interface{}) {
-	c.backend.EventMux().Post(ev)
-}
-
-func (c *core) handleMsg(payload []byte) error {
-	logger := c.logger.New()
-
-	// Decode message and check its signature
-	msg := new(message)
-	if err := msg.FromPayload(payload, c.validateFn); err != nil {
-		logger.Error("Failed to decode message from payload", "err", err)
-		return err
-	}
-
-	// Only accept message if the address is valid
-	// TODO: the check is already made in c.validateFn
-	_, src := c.valSet.GetByAddress(msg.Address)
-	if src == nil {
-		logger.Error("Invalid address in message", "msg", msg)
-		return tendermint.ErrUnauthorizedAddress
-	}
-
-	return c.handleCheckedMsg(msg, src)
-}
-
-func (c *core) handleCheckedMsg(msg *message, src tendermint.Validator) error {
-	logger := c.logger.New("address", c.address, "from", src)
-
-	// Store the message if it's a future message
-	testBacklog := func(err error) error {
-		// We want to store only future messages in backlog
-		if err == errFutureRoundMessage {
-			//We cannont move to a round in a new height without receiving a new block
-			var msgRound int64
-			if msg.Code == msgProposal {
-				var p tendermint.Proposal
-				if e := msg.Decode(p); e != nil {
-					return errFailedDecodeProposal
-				}
-				msgRound = p.Round.Int64()
-
-			} else {
-				var v tendermint.Vote
-				if e := msg.Decode(v); e != nil {
-					// TODO: introduce new error: errFailedDecodeVote
-					return errFailedDecodePrecommit
-				}
-				msgRound = v.Round.Int64()
-			}
-
-			c.futureRoundsChange[msgRound] = c.futureRoundsChange[msgRound] + 1
-			totalFutureRoundMessages := c.futureRoundsChange[msgRound]
-
-			if totalFutureRoundMessages >= int64(c.valSet.F()) {
-				c.startRound(big.NewInt(msgRound))
-			}
-
-		}
-		if err == errFutureHeightMessage || err == errFutureRoundMessage {
-			c.storeBacklog(msg, src)
-		}
-
-		return err
-	}
-
-	// TODO: check step before calling the relevant handler
-	switch msg.Code {
-	case msgProposal:
-		return testBacklog(c.handleProposal(msg, src))
-	case msgPrevote:
-		return testBacklog(c.handlePrevote(msg, src))
-	case msgPrecommit:
-		return testBacklog(c.handlePrecommit(msg, src))
-	default:
-		logger.Error("Invalid message", "msg", msg)
-	}
-
-	return errInvalidMessage
-}
-
-// TODO: re-implement to incorporate all three timeouts
-func (c *core) handleTimeoutMsg() {
-}
-
-//---------------------------------------Backlog---------------------------------------
-
-type backlogEvent struct {
-	src tendermint.Validator
-	msg *message
-}
-
-// checkMessage checks the message step
-// return errInvalidMessage if the message is invalid
-// return errFutureHeightMessage if the message view is larger than currentRoundState view
-// return errOldHeightMessage if the message view is smaller than currentRoundState view
-func (c *core) checkMessage(msgCode uint64, round *big.Int, height *big.Int) error {
-	if height == nil || round == nil {
-		return errInvalidMessage
-	}
-
-	if height.Cmp(c.currentRoundState.Height()) > 0 {
-		return errFutureHeightMessage
-	} else if height.Cmp(c.currentRoundState.Height()) < 0 {
-		return errOldHeightMessage
-	} else if round.Cmp(c.currentRoundState.Round()) > 0 {
-		return errFutureRoundMessage
-	} else if round.Cmp(c.currentRoundState.Round()) < 0 {
-		return errOldRoundMessage
-	}
-
-	// StepAcceptProposal only accepts msgProposal
-	// other messages are future messages
-	if c.step == StepAcceptProposal {
-		if msgCode > msgProposal {
-			return errFutureHeightMessage
-		}
-		return nil
-	}
-
-	// For steps(StepProposeDone, StepPrevoteDone, StepPrecommitDone),
-	// can accept all message types if processing with same view
-	return nil
-}
-
-func (c *core) storeBacklog(msg *message, src tendermint.Validator) {
-	logger := c.logger.New("from", src, "step", c.step)
-
-	if src.Address() == c.Address() {
-		logger.Warn("Backlog from self")
-		return
-	}
-
-	logger.Trace("Store future message")
-
-	c.backlogsMu.Lock()
-	defer c.backlogsMu.Unlock()
-
-	backlog := c.backlogs[src]
-	if backlog == nil {
-		backlog = prque.New()
-	}
-	switch msg.Code {
-	case msgProposal:
-		var p *tendermint.Proposal
-		err := msg.Decode(&p)
-		if err == nil {
-			backlog.Push(msg, toPriority(msg.Code, p.Round, p.Height))
-		}
-		// for msgRoundChange, msgPrevote and msgPrecommit cases
-	default:
-		var p *tendermint.Vote
-		err := msg.Decode(&p)
-		if err == nil {
-			backlog.Push(msg, toPriority(msg.Code, p.Round, p.Height))
-		}
-	}
-	c.backlogs[src] = backlog
-}
-
-func (c *core) processBacklog() {
-	c.backlogsMu.Lock()
-	defer c.backlogsMu.Unlock()
-
-	for src, backlog := range c.backlogs {
-		if backlog == nil {
-			continue
-		}
-
-		logger := c.logger.New("from", src, "step", c.step)
-		var isFuture bool
-
-		// We stop processing if
-		//   1. backlog is empty
-		//   2. The first message in queue is a future message
-		for !(backlog.Empty() || isFuture) {
-			m, prio := backlog.Pop()
-			msg := m.(*message)
-			var round, height *big.Int
-			switch msg.Code {
-			case msgProposal:
-				var m *tendermint.Proposal
-				err := msg.Decode(&m)
-				if err == nil {
-					round, height = m.Round, m.Height
-				}
-				// for msgRoundChange, msgPrevote and msgPrecommit cases
-			default:
-				var sub *tendermint.Vote
-				err := msg.Decode(&sub)
-				if err == nil {
-					round, height = sub.Round, sub.Height
-				}
-			}
-			if round == nil || height == nil {
-				logger.Debug("Nil view", "msg", msg)
-				continue
-			}
-			// Push back if it's a future message
-			err := c.checkMessage(msg.Code, round, height)
-			if err != nil {
-				if err == errFutureHeightMessage {
-					logger.Trace("Stop processing backlog", "msg", msg)
-					backlog.Push(msg, prio)
-					isFuture = true
-					break
-				}
-				logger.Trace("Skip the backlog event", "msg", msg, "err", err)
-				continue
-			}
-			logger.Trace("Post backlog event", "msg", msg)
-
-			go c.sendEvent(backlogEvent{
-				src: src,
-				msg: msg,
-			})
-		}
-	}
-}
-
-func toPriority(msgCode uint64, r *big.Int, h *big.Int) float32 {
-	// FIXME: round will be reset as 0 while new height
-	// 10 * Round limits the range of message code is from 0 to 9
-	// 1000 * Height limits the range of round is from 0 to 99
-	return -float32(h.Uint64()*1000 + r.Uint64()*10 + uint64(msgPriority[msgCode]))
-}
-
-//---------------------------------------NewUnminedBlock---------------------------------------
-
-func (c *core) handleUnminedBlock(unminedBlock *types.Block) error {
-	logger := c.logger.New("step", c.step, "height", c.currentRoundState.height)
-
-	if err := c.checkUnminedBlockMsg(unminedBlock); err != nil {
-		if err == errInvalidMessage {
-			logger.Warn("invalid unminedBlock")
-			return err
-		}
-		logger.Warn("unexpected unminedBlock", "err", err, "number", unminedBlock.Number(), "hash", unminedBlock.Hash())
-		return err
-	}
-
-	logger.Trace("handleUnminedBlock", "number", unminedBlock.Number(), "hash", unminedBlock.Hash())
-
-	c.latestPendingUnminedBlock = unminedBlock
-	// TODO: remove, we should not be sending a proposal from handleUnminedBlock
-	if c.step == StepAcceptProposal {
-		c.sendProposal(unminedBlock)
-	}
-	return nil
-}
-
-// check request step
-// return errInvalidMessage if the message is invalid
-// return errFutureHeightMessage if the height of proposal is larger than currentRoundState height
-// return errOldHeightMessage if the height of proposal is smaller than currentRoundState height
-func (c *core) checkUnminedBlockMsg(unminedBlock *types.Block) error {
-	if unminedBlock == nil {
-		return errInvalidMessage
-	}
-
-	//TODO: make the err message more specific to block maybe use consensus.ErrFutureBlock?
-	if c := c.currentRoundState.height.Cmp(unminedBlock.Number()); c > 0 {
-		return errOldHeightMessage
-	} else if c < 0 {
-		return errFutureHeightMessage
-	} else {
-		return nil
-	}
-}
-
-func (c *core) storeUnminedBlockMsg(unminedBlock *types.Block) {
-	logger := c.logger.New("step", c.step)
-
-	logger.Trace("Store future unminedBlock", "number", unminedBlock.Number(), "hash", unminedBlock.Hash())
-
-	c.pendingUnminedBlocksMu.Lock()
-	defer c.pendingUnminedBlocksMu.Unlock()
-
-	c.pendingUnminedBlocks.Push(unminedBlock, float32(-unminedBlock.Number().Int64()))
-}
-
-func (c *core) processPendingRequests() {
-	c.pendingUnminedBlocksMu.Lock()
-	defer c.pendingUnminedBlocksMu.Unlock()
-
-	for !(c.pendingUnminedBlocks.Empty()) {
-		m, prio := c.pendingUnminedBlocks.Pop()
-		ub, ok := m.(*types.Block)
-		if !ok {
-			c.logger.Warn("Malformed request, skip", "msg", m)
-			continue
-		}
-		// Push back if it's a future message
-		err := c.checkUnminedBlockMsg(ub)
-		if err != nil {
-			if err == errFutureHeightMessage {
-				c.logger.Trace("Stop processing request", "number", ub.Number(), "hash", ub.Hash())
-				c.pendingUnminedBlocks.Push(m, prio)
-				break
-			}
-			c.logger.Trace("Skip the pending request", "number", ub.Number(), "hash", ub.Hash(), "err", err)
-			continue
-		}
-		c.logger.Trace("Post pending request", "number", ub.Number(), "hash", ub.Hash())
-
-		go c.sendEvent(tendermint.NewUnminedBlockEvent{
-			NewUnminedBlock: *ub,
-		})
-	}
-}
-
-//---------------------------------------Propose---------------------------------------
-
-// TODO: add new message struct for proposal (proposalMessage) and determine how to rlp encode them especially nil
-// TODO: add new message for vote (prevote and precommit) and determine how to rlp encode them especially nil
-func (c *core) sendProposal(p *types.Block) {
-	logger := c.logger.New("step", c.step)
-
-	// If I'm the proposer and I have the same height with the proposal
-	if c.currentRoundState.Height().Cmp(p.Number()) == 0 && c.isProposer() && !c.sentProposal {
-		r, h, vr := c.currentRoundState.Round(), c.currentRoundState.Height(), c.validRound
-		proposal, err := Encode(&tendermint.Proposal{
-			Round:         r,
-			Height:        h,
-			ValidRound:    vr,
-			ProposalBlock: *p,
-		})
-		if err != nil {
-			logger.Error("Failed to encode", "Round", r, "Height", h, "ValidRound", vr)
-			return
-		}
-		c.sentProposal = true
-		c.backend.SetProposedBlockHash(p.Hash())
-		c.broadcast(&message{
-			Code: msgProposal,
-			Msg:  proposal,
-		})
-	}
-}
-
-func (c *core) handleProposal(msg *message, src tendermint.Validator) error {
-	logger := c.logger.New("from", src, "step", c.step)
-
-	var proposal *tendermint.Proposal
-	err := msg.Decode(&proposal)
-	if err != nil {
-		return errFailedDecodeProposal
-	}
-
-	// Ensure we have the same view with the Proposal message
-	// If it is old message, see if we need to broadcast COMMIT
-	//TODO: fixup
-	if err := c.checkMessage(msgProposal, proposal.Round, proposal.Height); err != nil {
-		if err == errOldHeightMessage || err == errOldRoundMessage {
-			// TODO: keeping it for the time being but rebroadcasting based on old messages should not be required due to partial synchrony assumption
-			// TODO: also need to add previous round messages to currentHeightRoundStates and only rebroadcast if older height
-			valSet := c.backend.Validators(proposal.ProposalBlock.Number().Uint64()).Copy()
-			previousProposer := c.backend.GetProposer(proposal.ProposalBlock.Number().Uint64() - 1)
-			valSet.CalcProposer(previousProposer, proposal.Round.Uint64())
-			// Broadcast COMMIT if it is an existing block
-			// 1. The proposer needs to be a proposer matches the given (Height + Round)
-			// 2. The given block must exist
-			if valSet.IsProposer(src.Address()) && c.backend.HasPropsal(proposal.ProposalBlock.Hash(), proposal.ProposalBlock.Number()) {
-				c.sendPrecommitForOldBlock(proposal.Round, proposal.Height, proposal.ProposalBlock.Hash())
-				return nil
-			}
-		}
-		return err
-	}
-
-	// Check if the message comes from currentRoundState proposer
-	if !c.valSet.IsProposer(src.Address()) {
-		logger.Warn("Ignore proposal messages from non-proposer")
-		return errNotFromProposer
-	}
-
-	// Verify the proposal we received
-	if duration, err := c.backend.Verify(proposal.ProposalBlock); err != nil {
-		logger.Warn("Failed to verify proposal", "err", err, "duration", duration)
-		// if it's a future block, we will handle it again after the duration
-		// TIME FIELD OF HEADER CHECKED HERE - NOT HEIGHT
-		if err == consensus.ErrFutureBlock {
-			c.stopFutureProposalTimer()
-			c.futureProposalTimer = time.AfterFunc(duration, func() {
-				c.sendEvent(backlogEvent{
-					src: src,
-					msg: msg,
-				})
-			})
-		}
-		return err
-	}
-
-	// Here is about to accept the Proposal
-	if c.step == StepAcceptProposal {
-		if c.proposeTimeout.started {
-			if stopped := c.proposeTimeout.stopTimer(); !stopped {
-				return errNilPrevoteSent
-			}
-		}
-		c.acceptProposal(proposal)
-		vr := proposal.ValidRound.Int64()
-		h := proposal.ProposalBlock.Hash()
-
-		if vr == -1 {
-			if c.lockedRound.Int64() == proposal.ValidRound.Int64() || h == c.lockedValue.Hash() {
-				c.sendPrevote(false)
-			} else {
-				c.sendPrevote(true)
-			}
-			c.setStep(StepProposeDone)
-		} else if rs, ok := c.currentHeightRoundsStates[vr]; vr > -1 && vr < c.currentRoundState.round.Int64() && ok && c.quorum(rs.Prevotes.VotesSize(h)) {
-			if c.lockedRound.Int64() <= proposal.ValidRound.Int64() || h == c.lockedValue.Hash() {
-				c.sendPrevote(false)
-			} else {
-				c.sendPrevote(true)
-			}
-			c.setStep(StepProposeDone)
-		} else {
-			return errNoMajority
-		}
-
-	}
-
-	return nil
-}
-
-func (c *core) acceptProposal(proposal *tendermint.Proposal) {
-	c.currentRoundState.SetProposal(proposal)
-}
-
-//---------------------------------------Prevote---------------------------------------
-
-func (c *core) sendPrevote(isNil bool) {
-	logger := c.logger.New("step", c.step)
-
-	var vote = &tendermint.Vote{
-		Round:  big.NewInt(c.currentRoundState.round.Int64()),
-		Height: big.NewInt(c.currentRoundState.Height().Int64()),
-	}
-
-	if isNil {
-		vote.ProposedBlockHash = common.Hash{}
-	} else {
-		vote.ProposedBlockHash = c.currentRoundState.Proposal().ProposalBlock.Hash()
-	}
-
-	encodedVote, err := Encode(vote)
-	if err != nil {
-		logger.Error("Failed to encode", "subject", vote)
-		return
-	}
-	c.broadcast(&message{
-		Code: msgPrevote,
-		Msg:  encodedVote,
-	})
-}
-
-// TODO: possibly need to add sentPrevote and sentPrecommit
-func (c *core) handlePrevote(msg *message, src tendermint.Validator) error {
-	// Decode PREPARE message
-	var prevote *tendermint.Vote
-	err := msg.Decode(&prevote)
-	if err != nil {
-		return errFailedDecodePrevote
-	}
-
-	if err = c.checkMessage(msgPrevote, prevote.Round, prevote.Height); err != nil {
-		// We want to store old round messages for future rounds since it is required for validRound
-		if err == errOldRoundMessage {
-			_, ok := c.currentHeightRoundsStates[prevote.Round.Int64()]
-			if !ok {
-				c.currentHeightRoundsStates[prevote.Round.Int64()] = *newRoundState(
-					big.NewInt(c.currentRoundState.Height().Int64()),
-					big.NewInt(prevote.Round.Int64()),
-					c.backend.HasBadProposal)
-			}
-
-			prevoteMS := c.currentHeightRoundsStates[prevote.Round.Int64()].Prevotes
-			if prevote.ProposedBlockHash == (common.Hash{}) {
-				prevoteMS.AddNilVote(*msg)
-			} else {
-				prevoteMS.AddVote(prevote.ProposedBlockHash, *msg)
-			}
-		}
-		return err
-	}
-
-	// Now we can add the prevote to our current round state
-	// TODO: manage prevote timer
-	if c.step >= StepProposeDone {
-		prevoteHash := prevote.ProposedBlockHash
-		curProposaleHash := c.currentRoundState.Proposal().ProposalBlock.Hash()
-
-		if prevoteHash == (common.Hash{}) {
-			c.currentRoundState.Prevotes.AddNilVote(*msg)
-		} else {
-			c.currentRoundState.Prevotes.AddVote(prevoteHash, *msg)
-		}
-
-		if c.step == StepProposeDone && !c.prevoteTimeout.started && c.quorum(c.currentRoundState.Prevotes.TotalSize(curProposaleHash)) {
-			c.prevoteTimeout.scheduleTimeout(timeoutPrevote(c.currentRoundState.Round().Int64()), c.currentRoundState.Height().Int64(), c.currentRoundState.Round().Int64(), c.onTimeoutPrevote)
-		} else if c.step == StepProposeDone && c.quorum(c.currentRoundState.Prevotes.NilVotesSize()) {
-			// TODO: probably need to stop timer, same in the other if branches need to fix this
-			c.sendPrecommit(true)
-			c.setStep(StepPrevoteDone)
-		} else if c.quorum(c.currentRoundState.Prevotes.VotesSize(curProposaleHash)) && !c.setValidRoundAndValue {
-			// this piece of code should only run once
-			if c.step == StepProposeDone {
-				c.lockedValue = &c.currentRoundState.Proposal().ProposalBlock
-				c.lockedRound = big.NewInt(c.currentRoundState.Round().Int64())
-				c.sendPrecommit(false)
-				c.setStep(StepPrevoteDone)
-			}
-			c.validValue = &c.currentRoundState.Proposal().ProposalBlock
-			c.validRound = big.NewInt(c.currentRoundState.Round().Int64())
-			c.setValidRoundAndValue = true
-		}
-	}
-
-	return nil
-}
-
-//---------------------------------------Precommit---------------------------------------
-func (c *core) sendPrecommit(isNil bool) {
-	logger := c.logger.New("step", c.step)
-
-	var vote = &tendermint.Vote{
-		Round:  big.NewInt(c.currentRoundState.round.Int64()),
-		Height: big.NewInt(c.currentRoundState.Height().Int64()),
-	}
-
-	if isNil {
-		vote.ProposedBlockHash = common.Hash{}
-	} else {
-		vote.ProposedBlockHash = c.currentRoundState.Proposal().ProposalBlock.Hash()
-	}
-
-	encodedVote, err := Encode(vote)
-	if err != nil {
-		logger.Error("Failed to encode", "subject", vote)
-		return
-	}
-	c.broadcast(&message{
-		Code: msgPrecommit,
-		Msg:  encodedVote,
-	})
-}
-
-func (c *core) sendPrecommitForOldBlock(r *big.Int, h *big.Int, digest common.Hash) {
-	sub := &tendermint.Vote{
-		Round:             r,
-		Height:            h,
-		ProposedBlockHash: digest,
-	}
-	c.broadcastPrecommit(sub)
-}
-
-func (c *core) broadcastPrecommit(sub *tendermint.Vote) {
-	logger := c.logger.New("step", c.step)
-
-	encodedSubject, err := Encode(sub)
-	if err != nil {
-		logger.Error("Failed to encode", "subject", sub)
-		return
-	}
-	c.broadcast(&message{
-		Code: msgPrecommit,
-		Msg:  encodedSubject,
-	})
-}
-
-func (c *core) handlePrecommit(msg *message, src tendermint.Validator) error {
-	// Decode COMMIT message
-	var precommit *tendermint.Vote
-	err := msg.Decode(&precommit)
-	if err != nil {
-		return errFailedDecodePrecommit
-	}
-
-	if err := c.checkMessage(msgPrecommit, precommit.Round, precommit.Height); err != nil {
-		// We don't care about old round messages, so if the message is in the correct height and round it should be fine
-		return err
-	}
-	// TODO: manage precommit timer
-	// We don't care about which step we are in to accept a precommit, since it has the highest importance
-	precommitHash := precommit.ProposedBlockHash
-	curProposaleHash := c.currentRoundState.Proposal().ProposalBlock.Hash()
-
-	if precommitHash == (common.Hash{}) {
-		c.currentRoundState.Precommits.AddNilVote(*msg)
-	} else {
-		c.currentRoundState.Precommits.AddVote(precommitHash, *msg)
-	}
-
-	if !c.precommitTimeout.started && c.quorum(c.currentRoundState.Precommits.NilVotesSize()) {
-		c.precommitTimeout.scheduleTimeout(timeoutPrecommit(c.currentRoundState.Round().Int64()), c.currentRoundState.Height().Int64(), c.currentRoundState.Round().Int64(), c.onTimeoutPrecommit)
-	} else if c.quorum(c.currentRoundState.Precommits.VotesSize(curProposaleHash)) {
-		c.commit()
-	}
-
-	return nil
-}
-
-//---------------------------------------Commit---------------------------------------
-func (c *core) handleCommit() {
-	c.logger.Trace("Received a final committed proposal", "step", c.step)
-	c.startRound(common.Big0)
 }
